requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
# created-by: conda 25.1.0
altair=5.0.1=py310hecd8cb5_1
arrow-cpp=19.0.0=hdba8a65_0
attrs=24.3.0=py310hecd8cb5_0
aws-c-auth=0.6.19=h6c40b1e_0
aws-c-cal=0.5.20=h3333b6a_0
aws-c-common=0.8.5=h6c40b1e_0
aws-c-compression=0.2.16=h6c40b1e_0
aws-c-event-stream=0.2.15=hcec6c5f_0
aws-c-http=0.6.25=h6c40b1e_0
aws-c-io=0.13.10=h6c40b1e_0
aws-c-mqtt=0.7.13=h6c40b1e_0
aws-c-s3=0.1.51=h6c40b1e_0
aws-c-sdkutils=0.1.6=h6c40b1e_0
aws-checksums=0.1.13=h6c40b1e_0
aws-crt-cpp=0.18.16=hcec6c5f_0
aws-sdk-cpp=1.11.212=h9475118_0
blas=1.0=mkl
blinker=1.9.0=py310hecd8cb5_0
boost-cpp=1.82.0=ha357a0b_2
bottleneck=1.4.2=py310hf3fd67a_0
brotli-python=1.0.9=py310h6d0c2b6_9
bzip2=1.0.8=h6c40b1e_6
c-ares=1.19.1=h6c40b1e_0
ca-certificates=2024.12.31=hecd8cb5_0
cachetools=5.5.1=py310hecd8cb5_0
certifi=2025.1.31=py310hecd8cb5_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py310hecd8cb5_0
cpuonly=2.0=0
filelock=3.17.0=pypi_0
freetype=2.12.1=hd8bbffd_0
fsspec=2025.2.0=pypi_0
gflags=2.2.2=hcec6c5f_1
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.43=py310hecd8cb5_0
glog=0.5.0=hcec6c5f_1
gmp=6.3.0=h6d0c2b6_0
gmpy2=2.2.1=py310h2cec913_0
icu=73.1=hcec6c5f_0
idna=3.7=py310hecd8cb5_0
intel-openmp=2023.1.0=ha357a0b_43548
jinja2=3.1.5=py310hecd8cb5_0
jpeg=9e=h46256e1_3
jsonschema=4.23.0=py310hecd8cb5_0
jsonschema-specifications=2023.7.1=py310hecd8cb5_0
krb5=1.20.1=h428f121_1
lcms2=2.16=h4f63f0c_0
lerc=4.0.0=h6d0c2b6_0
libabseil=20240116.2=cxx17_h548131f_0
libboost=1.82.0=hf53b9f2_2
libbrotlicommon=1.0.9=h46256e1_9
libbrotlidec=1.0.9=h46256e1_9
libbrotlienc=1.0.9=h46256e1_9
libcurl=8.11.1=h9bcc28a_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.22=h46256e1_0
libedit=3.1.20230828=h6c40b1e_0
libev=4.33=h9ed2024_1
libevent=2.1.12=h04015c4_1
libffi=3.4.4=hecd8cb5_1
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libgrpc=1.62.2=hf2926fd_0
libiconv=1.16=h6c40b1e_3
libjpeg-turbo=2.0.0=hca72f7f_0
libnghttp2=1.57.0=h9beae6a_0
libopenblas=0.3.21=h54e7dc3_0
libpng=1.6.39=h6c40b1e_0
libprotobuf=4.25.3=h34eed0b_0
libssh2=1.11.1=h3a17b82_0
libthrift=0.15.0=h70b4b81_2
libtiff=4.5.1=h6fa9cd1_1
libwebp-base=1.3.2=h46256e1_1
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_1
markdown-it-py=2.2.0=py310hecd8cb5_1
markupsafe=3.0.2=py310h46256e1_0
mdurl=0.1.0=py310hecd8cb5_0
mkl=2023.1.0=h8e150cf_43560
mkl-service=2.4.0=py310h46256e1_2
mkl_fft=1.3.8=py310h6c40b1e_0
mkl_random=1.2.4=py310ha357a0b_0
mpc=1.3.1=h46256e1_0
mpfr=4.2.1=h46256e1_0
mpmath=1.3.0=py310hecd8cb5_0
ncurses=6.4=hcec6c5f_0
networkx=3.4.2=py310hecd8cb5_0
numexpr=2.8.7=py310h827a554_0
numpy=1.26.4=py310h827a554_0
numpy-base=1.26.4=py310ha186be2_0
openjpeg=2.5.2=hbf2204d_0
openssl=3.0.15=h46256e1_0
orc=2.0.1=h5747287_0
packaging=24.2=py310hecd8cb5_0
pandas=2.2.3=py310h6d0c2b6_0
pillow=11.1.0=py310h47bf62f_0
pip=25.0=py310hecd8cb5_0
protobuf=4.25.3=py310h149b2d4_1
pyarrow=19.0.0=py310h6d0c2b6_1
pydeck=0.9.1=py310hecd8cb5_0
pygments=2.15.1=py310hecd8cb5_1
pysocks=1.7.1=py310hecd8cb5_0
python=3.10.16=hce00570_1
python-dateutil=2.9.0post0=py310hecd8cb5_2
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.2.2=py3.10_0
pytorch-mutex=1.0=cpu
pytz=2024.1=py310hecd8cb5_0
pyyaml=6.0.2=py310h46256e1_0
re2=2022.04.01=he9d5cce_0
readline=8.2=hca72f7f_0
referencing=0.30.2=py310hecd8cb5_0
requests=2.32.3=py310hecd8cb5_1
rich=13.9.4=py310hecd8cb5_0
rpds-py=0.22.3=py310h83de92b_0
setuptools=75.8.0=py310hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.2.1=h6d0c2b6_0
sqlite=3.45.3=h6c40b1e_0
streamlit=1.41.1=py310hecd8cb5_0
sympy=1.13.3=py310hecd8cb5_1
tbb=2021.8.0=ha357a0b_0
tenacity=9.0.0=py310hecd8cb5_0
tk=8.6.14=h4d00af3_0
toml=0.10.2=pyhd3eb1b0_0
toolz=1.0.0=py310hecd8cb5_0
torchaudio=2.2.2=py310_cpu
torchvision=0.17.2=py310_cpu
tornado=6.4.2=py310h46256e1_0
typing-extensions=4.12.2=py310hecd8cb5_0
typing_extensions=4.12.2=py310hecd8cb5_0
tzdata=2025a=h04d1e81_0
urllib3=2.3.0=py310hecd8cb5_0
utf8proc=2.6.1=h6c40b1e_1
wheel=0.45.1=py310hecd8cb5_0
xz=5.4.6=h6c40b1e_1
yaml=0.2.5=haf1e3a3_0
zlib=1.2.13=h4b97444_1
zstd=1.5.6=h138b38a_0
